fix: combine deposit source toenas and vote into a single transaction
